feat(plugin_runner): shared runtime engine
